dnsdist: Add an optional `status` parameter to `setAuto()`
If the parameter is present, its value is used to set the initial
state of the DownstreamServer until a health check is performed,
instead of using the previous known state.
It's useful if one wants to restore the automatic status detection
but knows that the server is either currently down but was previously
up (queries sent until the health check is performed will be lost);
or currently up but was previously down, so it process queries immediately.