Fix finding parent zone for DS queries
DS records must be answered from the parent zone (the zone
which has the delegation). Therefore, in case of DS queries,
PowerDNS does not use the first found authority but uses the
second one. This is correct if PowerDNS is authoritative
for the zone in the query, but will fail if PowerDNS is not
authoritative for the queried zone.

If PowerDNS is not authoritative for the queried zone, the first
found authority is already the correct one.

This patch also makes the respective code more readable.