dnsdist: Sort the servers based on their 'order' after it has been set
We keep the servers ordered inside their pools because it's easier
for the policies that way, so we sort them whenever a new one is
added. However we were doing the sorting _before_ the order of the
new server had been set, resulting in the last added server to be
sorted based on an order of 0, regardless of its actual order.
Reported by Frank Even (thanks!).

(cherry picked from commit 8c82c02db52398907377ea02f84527e5f29158ee)