rec: Ensure we don't hammer the RPZ master server
Before, if the `refresh` was not set in the lua-config file's
`rpzMaster` statement, we would keep trying to get delta's the whole
time. This commit ensures we update the zone's refresh config to the
value from the AXFR'd zone (if not set in the config).

(cherry picked from commit 2c68abe75d908b4a32a7de229c75af137dc3ec5a)