Add question for to increase-soa after non SOA edit
First stab at 3814.
Simple implementation that does simple SOA change detection. Should work for most people
who just want to let increase-serial do the work.

Logic to increase serial is not centralized and it does not detect if the SERIAL has changed in the SOA. It just looks for a change of the SOA record in general.