rec: Add an optional stack cache to MTasker
The stack cache is disabled by default and can be enabled by setting
`mtasker-stacks-cache-size` to a non-zero value. When the cache is
enabled the internal vector is not used, and a new one is
dynamically allocated instead, to be able to store it into a
per-MTasker free list when the context is released and then reuse
it for a new context later.