add more variables to carbon keys
The carbon key had the option to overwrite the hostname with a custom
one. But it was not possible to change either the instance or namespace,
which makes it harder to use in environments where multiple powerdns
instances are used for different purposes.

With this commit it is now possible to change the namespace and instance
part of the carbon key string. This makes it possible to differentiate
at more points or even group them easier together with other metrics,
for example the database used by pdns.