auth: geoip, add regression test for issue #7051
The added test does not validate that the weight system is working in
a pseudo-random way. It use a single entry per QType to ensure that
they are all queryable individually or via a ANY question.

The ANY test is done using TCP to avoid test failure due to exceeded
buffer size.