Use mmap for stack allocation, adding the MAP_STACK flag on OpenBSD.
On a userland to kernel transition OpenBSD checks if the stack
pointer points to a piece of mem marked MAP_STACK to prevent a large
class of exploits. Since we setup out own stack we must use mmap
on OpenBSD, and we might do that on other systems as well.