docs: Note that supermaster support is off by default since 4.2
With commit b80139777ebd2e0dc14cfd88d30eed80cfc04054 / PR #6021, it appears
that the behaviour of PowerDNS has changed to not act as a supermaster any
longer by default. Having to add a new setting that did not exist before,
to retain the old behaviour should be mentioned, I believe.