Prevent unlimited memory usage by cache
In my test-environment the cache would cause out of memory while max-cache-entries seemed to be ignored.
I first tried actually respecting max-cache-entries by not writing to the cache once it was more than full, but this seemed to cause a big penalty (in one test going from 17k QPS to about 11k QPS). 
The problem however seems to be reasonably under control with this change: if we don't skip cleanup if the cache is already too full. 

This change does still make things slower (now at around 14k QPS on my test-system when I try to trigger the out-of-memory), but the alternative was everything completely failing because of possible out of memory.
If s_maxcleaninterval is correctly chosen, then I expect the extra code to only be executed if we might be in the situation where the cache is being filled quicker than expected (and an out of memory is therefore a possibility).