Respect max-packet-cache-entries and max-cache-entries
Fixed proposed for #7458
This might replace or be combined with #7793 
This solution however seems to carry a smaller speed penalty and is less likely to be triggered when it's not actually needed.

If the initial attempt to clean the cache by only throwing away the old stuff wasn't sufficient to get below maxCached, then we will start throwing away from the top until we are at or close to the maximum. This way users are able to prevent out of memory by settings max-packet-cache-entries and max-cache-entries to appropriate values.

Code will execute only when max-cache has actually been exceeded. Overhead should be limited to when it's actually needed. 

I used dnsperf to test the speed and to stress-test on a 1G machine with LMDB backend and using fully random subdomains. Without this patch (and without #7793), PDNS would run out of memory after about 60 seconds. With the patch and with max-packet-cache-entries and max-cache-entries set to the default of 1000000 it would remain well within the memory limits.

My stress-tests triggered the code execution and even then speed penalty seems limited:
Two 60 second runs prior to the patch (runs cancelled just before it would otherwise go out of memory): 16981QPS and 17133QPS reported by dnsperf
Two 60 second runs with the patch: 16340QPS and 16988QPS