Fix SERVFAIL when backend returns empty DNSName
According to the documentation on UeberBackend::getAuth(), a backend
returning an empty DNSName should signal that no matching parent zone
exists in this backend. However commit ae14c1f36a raises an exception if
isPartOf() is called on an empty DNSName so we need to switch the order
of the logic to prevent a SERVFAIL as a result.

(cherry picked from commit 41147d678b30612d2431b59a898879368f1c4b06)