UeberBackend optimization
When all backends have returned a result then stop the loop. More modern
backends may be able to return a shorter- or empty DNSName so even with
the optimization to prevent re-querying on that specific backend, it
doesn't make sense to keep looping and checking caches until the shorter
query name reaches the specific qname that was returned.