geoipbackend: propagate weighted rounding gap fix
618824cb73db384d52cf379c7d49744dbb5b8f6d (PR#7219) fixed computation
of weighted round robin computation per Qtype.

That commit also introduced a bug on the removal of the rounding gap.

Because it works on a copy of a GeoIPDNSResourceRecord, the updated
weight of the last record is not propagated to the domain object.

The bug is as follow:
  1. The total weight is < 1000
  2. The random probabability gives 1000
  3. There is no resource to be given

This commit restore the original behavior of ensuring the gap is added
to the last resource so we always have 1000.