dnsdist: Wait longer for the TLS ticket to arrive in our tests
In TLS 1.3 the server sends the TLS ticket after the handshake has
been completed, but not necessarily right after that.
Ideally we would like to wait for up to several seconds, but stop
waiting as soon as we receive a ticket. Unfortunately we can't ask
that from `openssl s_client`, and we are currently not always waiting
long enough to get a ticket, leading to spurious failures.
Let's try waiting a bit longer to see if that helps.