rec: Fix IPv6-outgoing-only zero-scope ECS test
The recursor uses the outgoing address family to set the ECS address to
a localhost address when a client indicates that it wants no ECS (by
setting the all-zero scope).

Now that we _actually_ do v6-only outgoing when requested by configuration,
we need the regression test auths to listen on v6.

This change to the regression tests makes the ROOT auth listen on
[::1]:53, adds the hints, and for the ECS test, starts an IPv6 ECS
echo responder on [::1]:53000 and uses that address for the
forward-zone.