Update dynlistener.cc
according to definition   
https://en.cppreference.com/w/c/io/fgets  

```
Reads at most count - 1 characters from the given file stream and stores them in the character array pointed to by str. Parsing stops if a newline character is found, in which case str will contain that newline character, or if end-of-file occurs. If bytes are read and no errors occur, writes a null character at the position immediately after the last character written to str.
```


and according to definition    
https://en.cppreference.com/w/cpp/container/vector/resize

```
1) additional default-inserted elements are appended
```

I propose to exclude dead code as required by "CWE-561: Dead Code"