Also record the value that caused a hit.
For triggers fomr rpz zones it makes sense to store them as listed there.
For hit values (names or IPs) it makes more sense to store them
in the regular string value and not list them as rpz trigger format.
e.g.: a trigger is listed

	24.0.2.0.192.rpz-ip.rpz.local.

A corresponding hit as

	192.0.2.1