Add unit test for matching on netmask.
Also: fix the test to remove an entire entry. That should only be done
if there are no custom records left after the cleanup.
Old code was too eager and removed the entry whenever there was only one
custom record left. We could be trying to  remove a non-matching one.