builder-support: Extend 'all' module support to docs/sdist/deb targets
The build.sh script can build 'all' modules in a single execution, but
not all of the Dockerfile templates supported this mode. Since this is useful
for manual testing of the build process, this patch adds 'all' support
to the 'docs' and 'sdist' templates, and the templates which produce
Debian-style packages.

Signed-off-by: Kevin P. Fleming <kevin@km6g.us>