Avoid accidental removal of generated files
When generated files (man pages, api-swagger files, etc.) are
in the tree because they were part of a source distribution
but were not locally generated, 'make clean' should not remove
them. Doing so results in a broken tree, as the Makefiles
expect the files to be present, and running 'configure' again
will introduce the additional dependencies required to
generate them.