dnsdist: Deprecate parameters to webserver(), add 'statsRequireAuthentication' parameter
This PR deprecates the use of additional parameters with `webserver()`,
as the syntax is confusing and could lead to believe that the parameters
are per-instance while they actually are global.
Also implements an additional 'statsRequireAuthentication' parameter
to allow scraping the statistics without any kind of authentication,
which is useful to Prometheus setups with dynamic service discovery.