rec: Add simple EDNS0 padding
This commit ensures we do ignore the _value_ of the padding provided by
the client. but:

Ignoring the value of the PADDING option is not enough, as the OPTION LENGTH
can change per-request. However, ignoring both the LENGTH and the value,
still makes the packet uncachable as the the total EDNS0 RDLEN still counts
those bytes (and is already hashed by the time we reach the code to hash
the PADDING option).